# src/main/java/org/example/testasks/api/dto/ErrorResponseDto.java
package org.example.testasks.api.dto;

import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.Setter;

@Getter
@Setter
@AllArgsConstructor
public class ErrorResponseDto {
    private String error;
    private String message;
}